fix: load env-file at start of flag parsing
